backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
openai==1.58.1
anthropic==0.40.0
python-dotenv==1.0.1
pydantic-settings==2.7.0
python-multipart==0.0.19
